requirements.txt
pyTelegramBotAPI==4.7.0
python-dotenv==0.21.0
requests==2.25.1
loguru == 0.6.0
